+++ packages/ndk/lib/entities.dart
@@ -30,4 +30,7 @@ export 'domain_layer/entities/tuple.dart';
 export 'domain_layer/entities/user_relay_list.dart';
 export 'domain_layer/entities/blossom_blobs.dart';
 export 'domain_layer/entities/account.dart';
+
+/// models
+export 'data_layer/models/nip_01_event_model.dart';
 export 'domain_layer/entities/ndk_file.dart';