+++ packages/isar/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 0.2.3-dev.8
+
+ - **FIX**: clean imports.
+ - **FIX**: blossom signing.
+ - **FIX**: websocket state check optimization.
+ - **FEAT**: add test suite in actual cache tests.
+
+## 0.2.3-dev.7
+
+ - Update a dependency to the latest release.
+
+## 0.2.3-dev.6
+
+ - **FIX**: clean imports.
+ - **FEAT**: log color, params.
+
 ## 0.2.3-dev.5
 
  - **FIX**: clean imports.