+++ packages/rust_verifier/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.4.2-dev.8
+
+ - **REFACTOR**: other packages.
+ - **FIX**: clean imports.
+ - **FIX**: remove hex package depandance.
+ - **FEAT**: add doc.
+
+## 0.4.2-dev.6
+
+ - **FIX**: clean imports.
+ - **FIX**: remove hex package depandance.
+
 ## 0.4.2-dev.5
 
  - **FIX**: clean imports.